+++ crates/cairo-profiler/src/trace_reader/function_trace_builder/function_stack_trace.rs
